+++ Backpack/Classes/Radii/Radii.h
@@ -16,26 +16,8 @@
  * limitations under the License.
  */
 
-#import <XCTest/XCTest.h>
+#ifndef __BACKPACK_RADII__
+    #define __BACKPACK_RADII__
 
-#import <Backpack/Spacing.h>
-
-NS_ASSUME_NONNULL_BEGIN
-@interface BPKSpacingTest : XCTestCase
-
-@end
-
-@implementation BPKSpacingTest
-
-- (void)testSpacingsHaveCorrectValues {
-    XCTAssertEqual(BPKSpacingSm, 4);
-    XCTAssertEqual(BPKSpacingMd, 8);
-    XCTAssertEqual(BPKSpacingBase, 16);
-    XCTAssertEqual(BPKSpacingLg, 24);
-    XCTAssertEqual(BPKSpacingXl, 32);
-    XCTAssertEqual(BPKSpacingXxl, 40);
-}
-
-@end
-
-NS_ASSUME_NONNULL_END
+    #import "BPKRadii.h"
+#endif