+++ gulpfile.js
@@ -31,7 +31,7 @@ const PATHS = {
   output: path.join(__dirname, 'Backpack', 'Classes'),
 };
 
-const TYPES = new Set(['color', 'font', 'spacing']);
+const TYPES = new Set(['color', 'font', 'spacing', 'radii']);
 const VALID_TEXT_STYLES = new Set(['xs', 'sm', 'base', 'lg', 'xl']);
 const VALID_SPACINGS = new Set(['sm', 'md', 'base', 'lg', 'xl', 'xxl']);
 const WEIGHT_MAP = {
@@ -71,6 +71,15 @@ const convertFontWeight = weightString => {
   }
 
   return weight;
+};
+
+const generatePrefixedConst = ({ type, name, value }) => {
+  const capitalize = input => input.charAt(0).toUpperCase() + input.slice(1);
+  return {
+    type,
+    name: `BPK${capitalize(name)}`,
+    value,
+  };
 };
 
 const parseTokens = tokensData => {
@@ -139,18 +148,27 @@ const parseTokens = tokensData => {
     .filter(({ name }) =>
       VALID_SPACINGS.has(name.replace('spacing', '').toLowerCase()),
     )
-    .map(({ name, value }) => {
-      const capitalize = input =>
-        input.charAt(0).toUpperCase() + input.slice(1);
-      return {
+    .map(({ name, value }) =>
+      generatePrefixedConst({
         type: 'spacing',
-        name: `BPK${capitalize(name)}`,
+        name,
         value,
-      };
-    })
+      }),
+    )
     .value();
 
-  return _.chain([...colors, ...fonts, ...spacings])
+  const radii = _.chain(tokensData.properties)
+    .filter(({ category }) => category === 'radii')
+    .map(({ name, value }) =>
+      generatePrefixedConst({
+        type: 'radii',
+        name,
+        value,
+      }),
+    )
+    .value();
+
+  return _.chain([...colors, ...fonts, ...spacings, ...radii])
     .groupBy(({ type }) => type)
     .value();
 };